+++ sentry-android-okhttp/src/main/java/io/sentry/android/okhttp/SentryOkHttpInterceptor.kt
@@ -26,7 +26,7 @@ class SentryOkHttpInterceptor(
         val method = request.method
 
         // read transaction from the bound scope
-        var span = hub.span?.startChild("http.client", "$method $url")
+        val span = hub.span?.startChild("http.client", "$method $url")
 
         var response: Response? = null
 
@@ -53,12 +53,8 @@ class SentryOkHttpInterceptor(
             }
             throw e
         } finally {
-            if (span != null) {
-                if (beforeSpan != null) {
-                    span = beforeSpan.execute(span, request, response)
-                }
-                span?.finish()
-            }
+            finishSpan(span, request, response)
+
             val breadcrumb = Breadcrumb.http(request.url.toString(), request.method, code)
             request.body?.contentLength().ifHasValidLength {
                 breadcrumb.setData("request_body_size", it)
@@ -67,6 +63,22 @@ class SentryOkHttpInterceptor(
                 breadcrumb.setData("response_body_size", it)
             }
             hub.addBreadcrumb(breadcrumb)
+        }
+    }
+
+    private fun finishSpan(span: ISpan?, request: Request, response: Response?) {
+        if (span != null) {
+            if (beforeSpan != null) {
+                val result = beforeSpan.execute(span, request, response)
+                if (result == null) {
+                    // span is dropped
+                    span.spanContext.sampled = false
+                } else {
+                    span.finish()
+                }
+            } else {
+                span.finish()
+            }
         }
     }