+++ sentry/src/main/java/io/sentry/protocol/SentryTransaction.java
@@ -44,7 +44,9 @@ public final class SentryTransaction extends SentryBaseEvent {
     this.timestamp = sentryTracer.getTimestamp();
     this.transaction = sentryTracer.getName();
     for (final Span span : sentryTracer.getChildren()) {
-      this.spans.add(new SentrySpan(span));
+      if (Boolean.TRUE.equals(span.isSampled())) {
+        this.spans.add(new SentrySpan(span));
+      }
     }
     final Contexts contexts = this.getContexts();
     for (final Map.Entry<String, Object> entry : sentryTracer.getContexts().entrySet()) {
@@ -64,6 +66,13 @@ public final class SentryTransaction extends SentryBaseEvent {
             tracerContext.getStatus()));
     for (final Map.Entry<String, String> tag : tracerContext.getTags().entrySet()) {
       this.setTag(tag.getKey(), tag.getValue());
+    }
+
+    final Map<String, Object> data = sentryTracer.getData();
+    if (data != null) {
+      for (final Map.Entry<String, Object> tag : data.entrySet()) {
+        this.setExtra(tag.getKey(), tag.getValue());
+      }
     }
   }