+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## Unreleased
+
 ## 5.2.0-beta.2
 
 * Bump AGP to 7.0.2 (#1650)