+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-* Fix: Check at runtime if AndroidX.Core is available (#)
+* Fix: Check at runtime if AndroidX.Core is available (#1718)
 
 ## 5.2.0-beta.2