+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Feat: Add "data" to spans (#1717)
+* Fix: Should not capture unfinished transaction (#1719)
 
 ## 5.2.0-beta.2